Added explicit support for iOS platform in project and podspec. Bumped version to 1.0.1.

+++ MAChineLearning/BagOfWords/MLWordDictionary.m
@@ -178,7 +178,7 @@
 	
 	NSArray *wordInfos= [_dictionary allValues];
 	for (MLWordInfo *wordInfo in wordInfos)
-		[descr appendFormat:@"\t'%@': %lu (%lu)\n", wordInfo.word, wordInfo.totalOccurrencies, wordInfo.documentOccurrencies];
+        [descr appendFormat:@"\t'%@': %lu (%lu)\n", wordInfo.word, (unsigned long) wordInfo.totalOccurrencies, wordInfo.documentOccurrencies];
 	
 	[descr appendString:@"}"];
